refactor(studio-gen-eval): folder src into cohesive modules
Address PR review comment "Please properly folder all the files": the flat
src/ dump is grouped into per-concern folders — domain/ (shared types +
constants), client/, golden-set/ (set + import), scorer/, orchestrator/,
report/ — each with its colocated test. The CLI entry (run.ts) and its
import-order pin (bootstrap.ts) stay at src/ root so the documented
`tsx internal/studio-gen-eval/src/run.ts` path and TOOL_DIR resolution are
unchanged. Imports rewired to the new layout; no runtime behavior change.
