+++ internal/studio-gen-eval/src/domain/constants.ts
