+++ internal/studio-gen-eval/src/domain/types.ts
